+++ .php-cs-fixer.dist.php
@@ -1,14 +1,15 @@
 <?php
 
-use PhpCsFixer\{
-    Config, Finder
-};
+declare(strict_types=1);
+
+use PhpCsFixer\Config;
+use PhpCsFixer\Finder;
 
 $finder = Finder::create()
                 ->in(__DIR__ . '/src')
                 ->name('*.php')
                 ->ignoreDotFiles(true)
-                ->exclude(['bin','views','out','translations', 'metadata'])
+                ->exclude(['bin', 'views', 'out', 'translations', 'metadata'])
                 ->notPath(['src/metadata.php'])
                 ->ignoreVCS(true);
 
@@ -27,23 +28,22 @@ return $config
                        'operators' => [
                            '='  => 'align',
                            '=>' => 'align',
-                       ]
+                       ],
                    ],
-                   'blank_line_after_opening_tag' => true,
-                   'line_ending'                  => true,
-                   'class_attributes_separation'  => false,
-                   'blank_line_before_statement'  => ['statements' => ['break', 'continue', 'declare', 'throw', 'try']],
-                   'concat_space'                 => ['spacing' => 'one'],
-                   'declare_strict_types'         => true,
-                   'increment_style'              => ['style' => 'post'],
-                   'no_superfluous_phpdoc_tags'   => false,
-                   'no_useless_else'              => true,
-                   'no_useless_return'            => true,
-                   'ordered_class_elements'       => true,
-                   'ordered_imports'              => ['imports_order' => ['class', 'function', 'const']],
-                   'strict_comparison'            => true,
-                   'yoda_style'                   => ['equal' => false, 'identical' => false, 'less_and_greater' => false],
+                   'blank_line_after_opening_tag'                     => true,
+                   'line_ending'                                      => true,
+                   'class_attributes_separation'                      => false,
+                   'blank_line_before_statement'                      => ['statements' => ['break', 'continue', 'declare', 'throw', 'try']],
+                   'concat_space'                                     => ['spacing' => 'one'],
+                   'declare_strict_types'                             => true,
+                   'increment_style'                                  => ['style' => 'post'],
+                   'no_superfluous_phpdoc_tags'                       => false,
+                   'no_useless_else'                                  => true,
+                   'no_useless_return'                                => true,
+                   'ordered_class_elements'                           => true,
+                   'ordered_imports'                                  => ['imports_order' => ['class', 'function', 'const']],
+                   'strict_comparison'                                => true,
+                   'yoda_style'                                       => ['equal' => false, 'identical' => false, 'less_and_greater' => false],
                    'nullable_type_declaration_for_default_null_value' => false,
                ])
     ->setFinder($finder);
-