+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## Unreleased
+- Add PHPStan
+
 ## [v5.0.0] - 2023.10.24
 ### BREAKING
 - IMPORTANT! Drop Shopware 6.4 compatibility