+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Unreleased
 ### Add
 - Add reporting for Shopware marketplace
+- Add PHPStan
 
 ## [v5.0.0] - 2023.10.24
 ### BREAKING