+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 ## Unreleased
+### Add
 - Add PHPStan
+### Change
+- Set FactFinder cookies as required for Cookie Consent Manager
 
 ## [v5.0.0] - 2023.10.24
 ### BREAKING