+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## Unreleased
+## [v4.3.2] - 2024.01.26
 ### Fix
 - Filter empty attributes for ff-communication component
 ### Change
@@ -401,8 +401,9 @@
 ## [v1.0.0] - 2021.06.29
 Initial module release. Includes Web Components v4.0.3
 
-[v4.3.0]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.0
+[v4.3.2]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.2
 [v4.3.1]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.1
+[v4.3.0]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.3.0
 [v4.2.8]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.2.8
 [v4.2.7]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.2.7
 [v4.2.6]: https://github.com/FACT-Finder-Web-Components/shopware6-plugin/releases/tag/v4.2.6